Fix docker build error
The current requirements.txt depends on httptools version 0.1.x , which
has been upgraded to version 0.2.0 in the latest python3.7-alpine3.8 docker
image. Because of that, the build process require re-compiling of the
httptools package, which in turn require gcc alpine package, which
pretty problematic in Alpine based container.

Because of that, I decided to migrate our Dockerfile to 'slim' based
image (which utilize debian) which doesn't have this problem. More info
could be found in the original image's author explanation :
https://github.com/tiangolo/uvicorn-gunicorn-fastapi-docker#-alpine-python-warning